requirements.txt
accelerate~=0.26.1
openai~=1.11.1
loguru~=0.7.2
google-generativeai~=0.3.2
transformers~=4.37.2
jsonlines~=4.0.0
torch~=2.2.0
typer~=0.9.0
python-dotenv~=1.0.1
anthropic~=0.21.3